booking-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
requests==2.31.0
python-consul==1.1.0
python-dotenv==1.0.0
gunicorn==21.2.0